--- src/common/enums/enums.ts
@@ -1,2 +1,0 @@
-export { NotificationType } from './notification/notification'
-export { HeadingType } from './heading/heading'
