--- src/common/enums/heading/heading.ts
@@ -1,1 +1,0 @@
-export { HeadingType } from './heading-type.enum'
