+++ src/common/enums/heading/index.ts
@@ -1,0 +1,1 @@
+export { default as HeadingType } from './HeadingType.enum'
