+++ src/common/enums/index.ts
@@ -1,0 +1,2 @@
+export { NotificationType } from './notification'
+export { HeadingType } from './heading'
