+++ src/common/enums/notification/index.ts
@@ -1,0 +1,1 @@
+export { default as NotificationType } from './NotificationType.enum'
