--- src/common/enums/notification/notification.ts
@@ -1,1 +1,0 @@
-export { NotificationType } from './notification-type.enum'
