--- src/components/common/common.ts
@@ -1,1 +1,0 @@
-export { default as Heading } from './heading/heading'
