--- src/components/common/heading/common/helper/helper.ts
@@ -1,1 +1,0 @@
-export { getStyles } from './get-styles/get-styles.helper'
