+++ src/components/heading/helpers/index.ts
@@ -1,0 +1,1 @@
+export { default as getStylesByType } from './getStylesByType'
