+++ src/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Heading } from './heading/Heading'
