+++ src/services/index.ts
@@ -1,0 +1,5 @@
+import Notification from './notification.service'
+
+const notification = new Notification()
+
+export { notification }
