--- src/services/services.ts
@@ -1,5 +1,0 @@
-import { Notification } from './notifications/notifications.service'
-
-const notification = new Notification()
-
-export { notification }
